Cleanup the code and fix old issues
 - closes 98
 - closes 180
