+++ src/OpenRasta.Tests.Unit/GitHubIssueAttribute.cs
@@ -1,0 +1,11 @@
+using NUnit.Framework;
+
+namespace UriTemplateTable_Specification
+{
+  public class GitHubIssueAttribute : TestAttribute
+  {
+    public GitHubIssueAttribute(int issueNumber)
+    {
+    }
+  }
+}
